Break flexbox tests into separate pages
