+++ test/MUXControlsAdhocApp/FlexboxPages/FlexboxDirectionPage.xaml.cs
@@ -1,0 +1,12 @@
+﻿using Windows.UI.Xaml.Controls;
+
+namespace MUXControlsAdhocApp.FlexboxPages
+{
+    public sealed partial class FlexboxDirectionPage : Page
+    {
+        public FlexboxDirectionPage()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
