+++ test/MUXControlsAdhocApp/FlexboxPages/FlexboxOrder.xaml.cs
@@ -1,0 +1,12 @@
+﻿using Windows.UI.Xaml.Controls;
+
+namespace MUXControlsAdhocApp.FlexboxPages
+{
+    public sealed partial class FlexboxOrder : Page
+    {
+        public FlexboxOrder()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
