+++ test/MUXControlsReleaseTest/XamlIslandsTestApp/UwpApp/FlexboxPages/FlexboxPage.xaml.cs
@@ -12,49 +12,33 @@ namespace MUXControlsAdhocApp.FlexboxPages
         public FlexboxPage()
         {
             this.InitializeComponent();
-        }
 
-        private void InvalidateButton_Click(object sender, RoutedEventArgs e)
-        {
-            _flexbox.InvalidateMeasure();
-        }
-
-        private void DirectionComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            _flexbox.Direction = EnumValueFromComboBox<FlexboxDirection>(sender);
-        }
-
-        private void JustifyContentComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            _flexbox.JustifyContent = EnumValueFromComboBox<FlexboxJustifyContent>(sender);
-        }
-
-        private void AlignItemsComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            _flexbox.AlignItems = EnumValueFromComboBox<FlexboxAlignItems>(sender);
-        }
-
-        private void AlignContentComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            _flexbox.AlignContent = EnumValueFromComboBox<FlexboxAlignContent>(sender);
-        }
-
-        private void AlignSelfComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            Flexbox.SetAlignSelf(_alignSelf, EnumValueFromComboBox<FlexboxAlignSelf>(sender));
-            // TODO: How does an attached child property invalidate layout?
-            _flexbox.InvalidateMeasure();
-        }
-
-        private T EnumValueFromComboBox<T>(object sender) where T : struct, IComparable
-        {
-            string selection = (string)((ComboBox)sender).SelectedItem;
-            return Enum.Parse<T>(selection);
+            _contentFrame.Navigate(typeof(FlexboxPlayground));
         }
 
         private void NavigationView_BackRequested(Microsoft.UI.Xaml.Controls.NavigationView sender, Microsoft.UI.Xaml.Controls.NavigationViewBackRequestedEventArgs args)
         {
             Frame.GoBack();
+        }
+
+        private void NavigationView_ItemInvoked(Microsoft.UI.Xaml.Controls.NavigationView sender, Microsoft.UI.Xaml.Controls.NavigationViewItemInvokedEventArgs args)
+        {
+            string tag = args.InvokedItemContainer.Tag.ToString();
+            switch (tag)
+            {
+                case "Home":
+                    _contentFrame.Navigate(typeof(FlexboxPlayground));
+                    break;
+
+                case "Order":
+                    _contentFrame.Navigate(typeof(FlexboxOrder));
+                    break;
+
+                case "Direction":
+                    _contentFrame.Navigate(typeof(FlexboxDirectionPage));
+                    break;
+            }
+            
         }
     }
 }
